Moved the build tree method from TreeParser class to Tree class.
Changed the constructor of the TaxonNode so that it does
not demand a TaxonNode as parent but the id of the parent.
In this way it can be handled more easily if the
parent Node is not already a node of the tree.
In this way the link to the parent node has to be done later on.
(Disscussion)

Added the addNode method in Tree
class. (Based on the changed constructor)
The method does now search for every new node if it has
a father and / or children and links the potentially
found children and parents to another.