Updated a method call due to the new TaxonNode constructor.
Added the updateRoot method to link the root to its
children after the tree has been built.

Added a class that helps the user to download
a certain file.
Currently only working under linux due to me not
being able to find a command to download via command
line in windows.

The download helper does only download a file from
a given url to a given installation path.
It does not provide any form of automization yet.